[verify] Stop SSL cipher madness, and add proper support for TLS 1.3.
It requires some tweaking as pyOpenSSL does not support the required methods out of the box.
The principle is that instead of forcing the certificate using a restricted list of cipher, we use the signature_algorithms extension which is design for that purpose.
The cipher list methods is not practicable in TLS1.3 as the cipher no longer specify the certificate type.